requirements.txt
fastapi
uvicorn
kubernetes
pydantic
python-dotenv


